make promts usable as meta data
